user_backend/requirements.txt
apns2==0.4.1
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.1
Automat==0.6.0
boto3==1.7.28
botocore==1.10.28
certifi==2018.4.16
cffi==1.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
daphne==1.4.2
dj-database-url==0.4.0
Django==2.0.5
django-oauth-toolkit==1.1.0
django-push-notifications==1.6.0
djangorestframework==3.8.2
docutils==0.14
h2==2.6.2
haikunator==1.0.1
hpack==3.0.0
http-ece==1.0.5
hyper==0.7.0
hyperframe==3.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
jmespath==0.9.3
jsonfield==2.0.2
msgpack-python==0.5.6
oauthlib==2.0.7
pbr==4.0.3
psycopg2==2.7.4
psycopg2-binary==2.7.4
py-vapid==1.4.0
pycparser==2.18
PyJWT==1.6.1
python-dateutil==2.7.3
pytz==2018.4
pywebpush==1.7.0
redis==2.10.6
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
stevedore==1.28.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
zope.interface==4.5.0
Twisted[tls,http2]